타일맵구조생성(cmarkup),  구조전체적수정
<변경사항>
→ 타일툴 이용하여 진행은 미흡
   - 완성이된 타이미지를 생성

→ 구조 수정
   - 전반적인 클래스 정리 및 함수 정리

<미흡사항>
→ 타일맵 메모리닉 발생
→ 객체지향적인 부분의 구조는 미흡
→ 구조적인 부분 더욱더 수정예정
→ 타일툴로 이미지 생성후 파싱 진행상황 미흡

+++ Engine_SOURCE/jkConstantBuffer.cpp
@@ -16,7 +16,7 @@ namespace jk::graphics
 
 	bool ConstantBuffer::Create(size_t size)
 	{
-		desc.ByteWidth = size;
+		desc.ByteWidth = (UINT)size;
 		desc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_CONSTANT_BUFFER;
 		desc.Usage = D3D11_USAGE::D3D11_USAGE_DYNAMIC;
 		desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
