+++ Engine_Source/jkTime.cpp
@@ -25,7 +25,7 @@ namespace jk
 	{
 		QueryPerformanceCounter(&mCurFrequency);
 
-		double differnceFrequency = mCurFrequency.QuadPart - mPrevFrequency.QuadPart;
+		double differnceFrequency = (double)mCurFrequency.QuadPart - (double)mPrevFrequency.QuadPart;
 
 		mDeltaTime = differnceFrequency / mCpuFrequency.QuadPart;
